AlchemyMenu is done
Reverse-engineering has confirmed that AlchemyMenu is working with shallow copies of the InventoryEntryData objects in the player's inventory: the copies use pointers to the original entries' BaseExtraLists. This means that it's possible for us to merge BaseExtraList lists (not a typo) when two or more InventoryEntryData objects from different containers refer to the same item type. That fixes the problems we were having with the initial proof-of-concept.