Crafting research
Stripped to-do notes and debug logging out of the AlchemyMenu patch.

After further research, I've managed to discover that SmithingMenu is only used for tempering, while ConstructibleObjectMenu is used for crafting. I've put together a correct class definition for BSTHashMap, and documented the difference between crafting and tempering recipe definitions.

Cleaned up the definition for TESDataHandler, and made it possible to index into the form array array by form type as is done in Bethesda's code. These changes should hopefully be backward-compatible with anything already using this class definition.

I've also written debug tool that can be used to dump the RTTI for any VTBL (though it can't check on its own whether a VTBL has RTTI). "RTTI" in the context of this tool refers to the full set of run-time type information that MSVC baked into TESV.exe; by contrast, SKSE and its DYNAMIC_CAST macro refer to std::type_info instances (which definitely do not constitute the full collection of run-time type information for a class) as "RTTI."