SmithingMenu is patched
SmithingMenu now relies on both the player's inventory and the inventories of nearby containers, when checking whether items can be tempered and when actually tempering items. As a consequence of this change, ConstructibleObjectMenu now also checks nearby containers when checking whether items can be crafted, but not yet when crafting items (which means that this DLL is not fit for release in its current state).

Both menus still only display the number of items in the player's inventory when displaying the "Requires" text below the item preview.

Reverse-engineering: split the inventory classes into their own file. Made fixes to BSTHashMap.