Requirements text for crafting is now updated
We now patch the code to generate the required items text shown for the smithing and crafting menus. We need this because that text is supposed to show the amount of each ingredient available for use in crafting.

Added a class definition for BSTArray, so that I can have fine-grained control over these arrays without having to rely on calling vanilla methods (which is... messy... when dealing with a class templated on structs of different sizes).

At this point, we support everything we need to. We could maybe go and do enchanting soul gems as well, but I don't see that having as much benefit as the other kinds of crafting.